handles array of ints instead of a single double

+++ src/main/java/io/github/uclarocketproject/daqd/json/ItemJson.java
@@ -3,6 +3,6 @@ package io.github.uclarocketproject.daqd.json;
 public class ItemJson {
     public String name;
     public String id;
-    public double offset;
-    public double scale;
+    public int offset;
+    public int scale;
 }
